+++ scripts/cloud/cloud_debug.py
@@ -19,13 +19,13 @@ for file in os.listdir("E:/path/to/folder") :
     cloud = volmdlr.cloud.PointCloud3D.from_stl("E:/path/to/folder/" + file)
     cloud_faces = cloud.subdescription_2d()
     
-    volum = volmdlr.core.VolumeModel(cloud_faces)
+    volum = volmdlr.model.VolumeModel(cloud_faces)
     print('len(cloud_faces)', len(cloud_faces))
     faces.extend(cloud_faces)
     # volum.save_to_file(file)
     print(file, len(faces))
     
-volum = volmdlr.core.VolumeModel(faces)
+volum = volmdlr.model.VolumeModel(faces)
 volum.babylonjs()
         
 ################### READ STL
@@ -33,6 +33,6 @@ volum.babylonjs()
 # cloud3d = volmdlr.cloud.PointCloud3D.from_stl(path)
 # faces = cloud3d.subdescription_2d(resolution=20)
 
-# volum = volmdlr.core.VolumeModel(faces)
+# volum = volmdlr.model.VolumeModel(faces)
 
 # volum.babylonjs()