+++ volmdlr/step.py
@@ -18,7 +18,7 @@ import dessia_common.core as dc  # isort: skip
 from dessia_common.files import BinaryFile  # isort: skip
 
 import volmdlr
-import volmdlr.core
+import volmdlr.model
 import volmdlr.edges
 import volmdlr.faces
 import volmdlr.curves
@@ -758,7 +758,7 @@ class Step(dc.DessiaObject):
             given class.
         :type show_times: bool
         :return: A volmdlr solid object.
-        :rtype: :class:`volmdlr.core.VolumeModel`
+        :rtype: :class:`volmdlr.model.VolumeModel`
         """
         object_dict = {}
         times = {}
@@ -793,7 +793,7 @@ class Step(dc.DessiaObject):
             print()
 
         if self.root_nodes["NEXT_ASSEMBLY_USAGE_OCCURRENCE"]:
-            return volmdlr.core.VolumeModel([self.instatiate_assembly(object_dict)])
+            return volmdlr.model.VolumeModel([self.instatiate_assembly(object_dict)])
         primitives = []
         shapes = [object_dict[shape] for shape in shape_representations
                   if self.functions[shape].name in STEP_REPRESENTATION_ENTITIES]
@@ -802,8 +802,8 @@ class Step(dc.DessiaObject):
                 primitives.extend(shape)
             else:
                 primitives.append(shape)
-        volume_model = volmdlr.core.VolumeModel(primitives)
-        # volume_model = volmdlr.core.VolumeModel([object_dict[shell_node] for shell_node in shell_nodes])
+        volume_model = volmdlr.model.VolumeModel(primitives)
+        # volume_model = volmdlr.model.VolumeModel([object_dict[shell_node] for shell_node in shell_nodes])
         return volume_model
 
     def _helper_instantiate(self, node, object_dict, times, show_times):